+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * Fixes handling naive datetimes in the admin's requests overview.
 * Fixed url() deprecation warnings for Django >= 3.1.
+* Add support for django >= 3.1 native JSONField